requirements.txt
numpy<2
torch
openai-whisper
tiktoken
ffmpeg-python
python-telegram-bot==13.15
psutil
